android notif [nfc]: Organize both notif-builder blocks, and add comments
The aim here is to put the two in the same order, so that it's easier
to compare them.  Also in a logical order to the extent that's
possible: so general to specific (starting with the group, and whether
it's a group summary), and earliest to latest (so putting at the end
the parts about what happens if the user opens the notification.)

Then add some comments that come to mind from comparing the two,
or from reading the docs.