android notif [nfc]: Rename file to NotificationUiManager
This makes a much more on-point name for what this file actually does.
It doesn't have anything particularly to do with FCM at all; that's
the job of some other code (and not very much of it.)  Rather its job
is to maintain the set of notifications as they appear in the UI.

The reason it was named this is just that that's what it had always
been; converted from FCMPushNotifications.java in 5654221db, and
before that renamed from GCMPushNotifications.java in f180171c2,
both early 2019.

And that in turn dates to when the first version of this code was
written in 2017, where it was entangled with a pretty chaotic
third-party library from the RN universe.  (See e.g. 30e4d8efa,
and the rest of the changes leading up to d048dba00.)