android notif: Call `launchMainActivity` in `notifyReact` unconditionally
The intent used to start the main activity has the
FLAG_ACTIVITY_NEW_TASK flag. The doc on that flag says this [1]:

> When using this flag, if a task is already running for the
> activity you are now starting, then a new activity will not be
> started; instead, the current task will simply be brought to the
> front of the screen with the state it was last in.

That should mean that there's no hiccup or loss of state if we run
our code to launch the main activity from the main activity when
it's already foregrounded. Empirically that seems to be the case.

I didn't find any history that says the conditional is necessary;
looks like it dates back to the old Wix library we used to use
here [2].

[1] https://developer.android.com/reference/android/content/Intent#FLAG_ACTIVITY_NEW_TASK
[2] https://github.com/zulip/zulip-mobile/pull/5146#discussion_r762337799