persist [nfc]: Rearrange persistStore a bit
This reorders the source code, but preserves the order things
actually happen in.  Hopefully it makes that control flow easier
to see: we create the persistor, set an immediate task to restore
to it, and promptly return.  All that other code runs later.

Also add a couple of blank lines to avoid making things look
grouped together that logically aren't.