checkout-keystore: Filter some annoying `sq` output, with fancy Bash
The output in question looks like this:

  $ tools/checkout-keystore
  Enter password to decrypt key Greg Price <greg@zulip.com> (1F94 DDB6 B44C 4D8D):
  Encrypted using AES with 256-bit key
  Compressed using ZIP

  Wrote decrypted file: /home/greg/z/mobile/android/release.keystore

  This file will be deleted in 15 minutes.

The "Encrypted using …" and "Compressed using …" lines are useless at
best, and can easily be confusing -- in context here, they read as if
they're saying the output file got compressed that way.  I'm not even
entirely sure what they do mean -- whether they're referring to the
input cryptotext, or to the key file.