actions [nfc]: Drop trivial API-wrapper setSubscriptionProperty
This is just a wrapper for api.setSubscriptionProperty, and one that
doesn't add any meaningful semantics.  That is, not only does it not
add any significant behavior or logic, but there also isn't any way
its semantics meaningfully differs from the underlying function.

(Sometimes a function with a trivial implementation is nevertheless
abstracting something away, so that the fact the implementation is
trivial is itself a nonobvious fact.  But this isn't one of those.)

So, just inline it at its call sites.  We already have a series of
other sites that call `api.setSubscriptionProperty` directly.

Originally noticed this here:
  https://github.com/zulip/zulip-mobile/pull/4899#pullrequestreview-772130836