android: Be manifestly lazy in getting React context
These lines where we get the React context look like they could be
cheap and side-effect-free… or could not be.  If not, it'd be good
to avoid running them when not needed.

I think there isn't any live case where it'd matter even if they are
expensive -- the only intents we handle other than the two cases
here (which do use the React context) are the web-auth intents,
where we'll want a React instance/context anyway -- but best not to
have to worry about that possibility coming up in the future.

To avoid duplicating our path through some details of React Native's
API, we factor it out as a couple of handy extension methods.