expo-sqlite tests: Write a mock version of expo-sqlite
Mostly this amounts to a fourth platform-specific implementation for
expo-sqlite: upstream has implementations for Android, iOS, and web,
and this is one for Node.

The one "mock" aspect is that this keeps the underlying databases in
memory, rather than on disk.  That's helpful for isolation.  There's
also an extra function provided, to delete a database.

Include a small smoke-test to demonstrate that this does indeed run.