expo-sqlite tests: Get a working `immediate` in Jest
The `@expo/websql` package (aka `node-websql`) that provides most of
the JS layer of expo-sqlite uses in turn the `immediate` package to
get an implementation of basically `setImmediate`.

In Jest tests, that doesn't work; the symptom is that the test hangs
and then times out at 5 seconds, not having completed.  Looking at the
`immediate` implementation, it relies on `process.nextTick` (when
that's present, i.e. on Node).  With Jest's "modern" fake timers --
which we now use, and have for a while -- that's mocked out, so that
explains why the "immediate" callbacks never get run.  (The
implementation also uses `setTimeout`, in what seems to be a
convoluted substitute for try/catch.  So that'd be an additional
reason it wouldn't work, if it got that far.)

We live in a JS world that has Promises.  Given those, setImmediate
has a trivial implementation.  Replace the whole module with that.

Quite likely we want a similar substitution in the actual app, but
leave that for another time for now, because in the app this module
seems to at least be working at all.