android: Update target SDK version to 31, i.e., Android 12
This change will be required in order to upload new releases to the
Play Store, effective probably 2022-11-01. That isn't very soon, but
we cut it kind of close last time, and we don't want to this time.

The change causes Android 12 and later to apply to our app a number
of behavior changes introduced in that version, documented here:
  https://developer.android.com/about/versions/12/behavior-changes-12

Most of the work for this was #5145, "android notif: Replace our
`NotificationIntentService` with directly opening MainActivity".
Earlier in this series of commits, we also fixed #5171 ("Handle
stricter length limit on Android toasts when targeting Android 12")
and handled the "Safer component exporting" change described at
  https://developer.android.com/about/versions/12/behavior-changes-12#exported .

There were a few other things that stood out in that "Behavior
Changes" article, but none that seem to require any specific action
from us; see
  https://github.com/zulip/zulip-mobile/issues/5101#issuecomment-996303224 .

Fixes: #5101