ZulipText: Rename, from RawLabel
Not NFC just because of the `invariant` change in the definition
file, which affects logging.

This component has long been appropriately used to render text that
isn't accurately described as a "label": data, titles, descriptions,
etc.

React Native's suggested convention for this kind of component is
fine [1]:

> The recommended way to use consistent fonts and sizes across your
> application is to create a component `MyAppText` that includes
> them and use this component across your app. You can also use this
> component to make more specific components like `MyAppHeaderText`
> for other kinds of text.

The `Label` component has had a role like the "more specific
components" in that last piece of advice, since 56819636a. It uses
this basic component, but adds translation as a special feature. The
name `Label` isn't great for that component, for the same reason
that `RawLabel` wasn't good for this one. We'll give it a better
name next.

[1] https://reactnative.dev/docs/text#limited-style-inheritance