UnreadNotice: Give better (but not yet perfect) translation for plural
Not yet perfect because some languages have categories for "two" or
"few", etc., as Greg points out; see
  https://github.com/zulip/zulip-mobile/pull/4893#discussion_r668406113

So, start an experiment to see if we can do better, using ICU
syntax.

The ICU syntax is likely to be a barrier to translators if they have
to deal with it directly. It's confusing, and more complicated than
the other strings they're used to working with.

Transifex has some documentation that suggests it might offer a
special UI that helps the user along, if ICU syntax is used
correctly, but we haven't seen such a UI in practice yet. So, add an
entry to messages_en.json that Transifex will hopefully like (but
don't use the message in the UI yet, just in case not). The entry
will be uploaded to Transifex next time we do a sync, and then we'll
see.