expo-sqlite tests: Demo the broken asynchrony handling
This test shows that if using expo-sqlite you try to make a
transaction that involves some asynchronous work between SQL queries,
the transaction gets committed in a half-done state.

This is a limitation that stems from Web SQL Database, the API (an
old proposed standard, implemented in Chromium-based browsers but no
others) that expo-sqlite mostly provides: that API doesn't provide
any way for the implementation to know when you're done with any
async work, so effectively it assumes you don't try to do any.

Moreover there's an implementation bug, which isn't due to the API:
when you do try to do this, the subsequent statements you try to add
to the transaction get silently ignored, with no error.

We'll work around this issue in the promisifying wrapper we'll add
atop expo-sqlite.  Meanwhile, this test serves to demonstrate what
the issue is in the first place.