narrow: Cut all unneeded calls to streamNameOfNarrow
In almost all the places we were using streamNameOfNarrow, what we
really want is just the stream ID, not its name; or an actual Stream
or Subscription object.  Now that the narrows contain the stream ID,
we can get it directly, and look up the Stream or Subscription by that
instead of by name.

The only exceptions are where we construct new Narrows from old ones,
either to go from a stream narrow to a topic narrow or vice versa.
After this commit, the only call sites of streamNameForNarrow are
those and its own tests.

This change should be NFC except in cases where the stream name and ID
don't agree -- e.g., because the stream has been renamed.  In those
cases, it fixes a bug.