messages: Fix several || that should be ??
The point of each of these is to fall back if the given property is
*absent*, not if it's present with a falsy value.

In the case of `submessages`, `edit_history`, or `subject_links`,
this is NFC because when present they're arrays, hence always truthy.

For `subject` this is NFC because the topic is never the empty string.
(When a client tries to send a message with empty topic, it gets
turned into the string "(no topic)".)

For `rendered_content` this may not be NFC; I'm not sure if the
content can ever be the empty string.  If so then this fixes our
behavior in that case, where we'd wrongly ignore an update that made
the content empty.