message: Handle the server's stealth edits for inline URL previews
This peculiarity is now documented, since zulip/zulip#20693.
See also chat discussion:
  https://chat.zulip.org/#narrow/stream/378-api-design/topic/.60update_message.60.20event/near/1299957

Once we make the types aware of this possibility, and then edit
our code as the type-checker points out we need to (by bailing
out of trying to construct a history entry when we don't have the
user_id and timestamp we'd need for it), it actually turns out
that giving them their correct behavior (i.e., suppressing them
from the history) is just about the easiest way we can handle it.
So just go ahead and do that, and hooray for type-checking.