unread [nfc]: More explicitly name orig/new values in update_message handler
This helps isolate the specifics of the event's format (with the
fallbacks from one property to another) and let the rest of this
logic be read more straightforwardly.

Especially good because that fallback logic turns out not be quite
the correct fallback logic!  We'll fix that shortly, and this helps
set us up for that.