action types: Mark remaining object types read-only
Previously we covered all the object types that are given names,
with `type Foo = …` declarations.  This covers the remainder,
found more deeply nested inside such types.

Found relevant types with:

  $ git grep -C2  '^ .*\{' src/{action,api/event}Types.js

and then just edited by hand.  We also take the opportunity to tighten
the formatting slightly in a couple of places.