persist: On error deserializing, don't store null
We don't use these nulls in any later handling; they go into the
REHYDRATE payload, but then in redux-persist-migrate we cast them
away to pretend they can't be there.  It's not clear how we would
usefully do anything with them, either.

Instead, on an error here just omit that key, the same as we would
if it weren't there in the underlying storage at all.

This lets us cut out a fixme step in redux-persist-migrate.  The type
of the REHYDRATE payload is still a lie, but it's now the same lie on
its way into this code as out of it.

In the one place our reducers look at this payload, update the type
and comment, but nothing substantive changes -- we were already
testing for the there-relevant property being falsy, which is just as
appropriate for "could be missing" as for "could be missing or null".