checkout-keystore [nfc]: Use less-fancy Bash for filtering stderr
Thanks to Anders for the suggestion:
  https://github.com/zulip/zulip-mobile/pull/5144#issuecomment-985778181

I wasn't sure if /dev/fd/N was available on macOS, but indeed it
seems to be, and this `>(…)` Bash syntax works just fine there.
(Even in the ancient /bin/bash that Apple ships.)

This sure is simpler to read!  Really not much different from
the usual case of applying a grep to a command's stdout.  More
shell-style, in the spirit of shell pipelines, while the `coproc`
approach feels lower-level.

I kept the grep command pulled out separate from the main `sq`
command, as a tiny shell function, in order to have a place to put
the comment explaining what it's there for.

Suggested-by: Anders Kaseorg <anders@zulip.com>