api types: Mark realm_user_groups as required, relying on Zulip 1.8
The one thing we were doing to accommodate it possibly being missing
is a `|| []` in the reducer.  Leave that in place, because it's such
a small amount of code.

In the future we might validate that the data we get from the server
actually satisfies the expected types.  But without some kind of
framework for that, rejecting a result that's missing this would take
more code than just falling back to `[]` as we do.