narrow: Use narrow's stream ID to get name in apiNarrowOfNarrow
We still need a stream name here for the actual request to the
server, until we start requiring server 2.1.  But we can look up
the name by ID in the overall state, just as we do for PM narrows.

In particular this means that we'll now behave correctly if the
stream got renamed since we formed this Narrow object (although
until we push IDs all the way through to the server, there'll
still be a race if the stream gets renamed concurrently with our
request.)