stream [nfc]: Add TODO-3918 to all remaining uses of stream names
Or in some cases, there's an existing TODO-server comment, and we
just add a reference to the issue.  (These are places where there's
nothing to be done about the issue before simply dropping support
for older servers, so the proper thing for the TODO itself to refer
to is the server version.)

I believe this covers all remaining instances of #3918.  That means
that every place we refer to a stream name should either:
 * be a legitimate use of stream names;
 * have a TODO-3918 comment; or
 * have a different kind of TODO comment (as it happens, always a
   TODO-server comment), and a mention of #3918.

As a result, a search like `git grep TODO..3918` will find all the
remaining places we need to act in order to close issue #3918,
according to the plan I described here:
  https://github.com/zulip/zulip-mobile/issues/3918#issuecomment-1026479673

The broader search `git grep -w 3918` will additionally find all the
places that are (known) instances of the issue but are blocked purely
on server upgrades.