unread [nfc]: Cut a bogus empty-string fallback on stream name
Our types say this is a string, so the `|| ''` can't have any
effect.

Just to be sure there wasn't a latent bug somewhere that this
was hiding, I went and traced where this data comes from.
That `streamName` property comes from these lines in
getUnreadStreamsAndTopics, in unreadSelectors:

      const { name, color, in_home_view, invite_only, pin_to_top } =
        subscriptionsById.get(streamId) || NULL_SUBSCRIPTION;

      // …
        streamName: name,

So it's the `name` property either of an actual subscription,
or of NULL_SUBSCRIPTION.

In an actual subscription that's a string, and for that matter
shouldn't be empty.  In NULL_SUBSCRIPTION it's… the empty string.
So effectively we're already doing this fallback up in
getUnreadStreamsAndTopics, where we construct this data in the
first place.

It's not clear that that's the right answer either, but let's at
least do the fallback just once.  We're already doing it
systematically there (for this and several other properties) and
haphazardly in this file (which already has another reference to
streamName without this fallback), so let it be there for now.