Revert "webview html: Get stream name from stream, rather than message"
This reverts most of commit 9d05bbaf6fc9937ac9a776e28cbc988cc3b47b45.
We leave the test changes in place, because those just set up better
fixture data.

It's possible for the message list to show messages in streams that
we don't actually have data for: in particular, streams that our user
was previously a member of (and received some messages there while
they were), but no longer is, and can't see (because it's private and
they're not an admin, or because they're a guest.)  As a result, we
can't simply get the stream name here from our streams data.

There are a couple of ways we could go to fix the issue that commit
originally fixed, so that recipient headers get updated when a stream
is renamed; I'll discuss those in or linked from the issue #5206.
In the meantime, revert, because a crash when you've left the stream
is a lot worse than an out-of-date name when it's been renamed.

Fixes: #5206