action types [nfc]: Get state type out of PresenceResponseAction
This isn't *quite* part of our API types -- not as much so as in the
case of event actions, where `eventToAction` systematically mirrors
the server API in the actions it creates.  Here, the action is created
by its own helper reportPresence, which gets this data from an API
response but could then freely choose what form of data to put into
the resulting action.

But still, the fact that these actions carry the same data structure
we use for the state we maintain is a bit of an accident of a design
detail.  If in the future our internal state data structure were to
diverge from what's in the API here, the most natural thing would be
for the action to continue to reflect the API response, not the state
-- it'd be the reducer's job to apply that data appropriately to the
new state data structure.

Between this, the recent commit removing AlertWordsState here, and
the one before that for MuteState, we now no longer have any types
here that use our internal data structures to describe the server API.