types: Stop sharing UserStatus with /register response type
There's no reason why the shape of the data coming from the server
has to be the same shape we use in our internal representations.

Now, it'll be easy to give different definitions for the relevant
part of the /register response and for what we put in Redux.

Since UserStatus is now basically just a helper for UserStatusState,
move it there alongside.