mute tests: Well-type muteReducer tests -- and fix complete nonsense
This is some legacy code that's been barely touched since 2017.
And it turns out... not a single one of these test cases has the
data in the form it actually appears in.  Or for that matter in a
form the data *could* possibly appear in in order for this feature
to work.

Happily, when we bring the type-checker to bear on this file, it
immediately tells us about that!  Do so, and fix the brokenness.