notif [nfc]: Add TODO comments for using the new stream_id
The repeated mentions of the issue may look redundant.  But this
way, when the TODO-3918 comments are deleted in the course of
taking care of them, the other references will naturally remain.

That will help with grepping for remaining places we're using
stream names, even while those places aren't directly actionable
because they're blocked on a minimum-server-version upgrade.