react-nav config [nfc]: Fold away never-varying parameters
Now that we're not trying to unify the presentation of bottom tabs
with that of top tabs, these always have the same constant value at
each call site.

Which is appropriate, because they're describing how the designs
of those two different UI components work.

So drop the parameters, and move the values into the respective
central configs.