stream [nfc]: Simplify out most of PseudoSubscription in StreamListCard
Here there is one piece of meaningful data in these objects beyond
what's in a Stream, so we can't drop this whole type immediately.
But getting rid of the union prepares the ground for clearing that
up too; and in fact this type isn't even the biggest improvement
we'll get when we do so.

The three other properties here are therefore always void, so
constant-fold them accordingly.

This also makes clear that there's only one section in this
SectionList, so that's something we can simplify away next too.