stream: Use verbatim `Stream`s in StreamListCard, for less garbage
This avoids constructing a bunch of new Stream-like objects each time
we render this component.

The change should be NFC apart from any performance effect.

This also opens up a further optimization opportunity: now that the
items we're working with here are verbatim Stream objects from our
state, we could memoize the step where we sort them.  Comment that.