api [nfc]: Factor out messageMoved to look at update_message events
We'll want to reuse this bit of logic in multiple sub-reducers, so
it's time it got factored out from this one.

And in fact putting it inside the API code seems most appropriate:
it's basically part of presenting an improved version of what the API
*could* have been, if designed from the outset with knowledge of all
the features that would be added.