deps: Upgrade react-intl to 5.24.6, the latest
I think we still don't trust the maintainers to have a strategy on
breaking changes (see f9a20cae9 and 803852839), so we'll still leave
out the caret here. But I've kicked the tires on iOS and Android and
this seems to be fine.

See changelog:
  https://github.com/formatjs/formatjs/blob/main/packages/react-intl/CHANGELOG.md
Nothing listed there looks likely to be breaking.