logging: Set server-version tags immediately on fetch
We've been updating these only after the reducers do their work.
But there's quite a bit of surface area in the reducers -- and
particularly so in the handling of REGISTER_COMPLETE actions.
So we do sometimes have bugs that cause exceptions there.

When we do, make sure we get accurate information about what
Zulip server versions might be involved in triggering them,
to help with debugging.