lint: Turn on import/no-cycles, forbidding runtime import cycles
We still have some large import cycle-blobs at type-checking time.
But the more severe kind of import cycle is those that survive in
the actual runtime code, so it's worth keeping these out of our
codebase.

A runtime import cycle can cause crashes, because it inevitably
means that one module's toplevel code must run while that of a
module it depends on hasn't yet.  This is benign so long as the
references are all inside function definitions that don't actually
get executed by the toplevel code of any of the modules in the
cycle (and therefore don't get executed until all the cycle's
modules have been properly initialized), and so I believe there's
no live bug that was fixed by the decycling in this branch leading
up to here.

But we have had bugs -- at least in development, not sure any have
ever made it as far as the main branch -- caused by import cycles.
And once there are some import cycles lying around, it becomes
easier and easier for the next stray import to make a larger and
less-benign cycle-blob, and to introduce a live issue.

As a checkpoint, here's the current state of our import cycle-blobs
at type-checking time.  They're much like they were earlier in this
branch, just after eliminating "common/index.js":
  43 files around src/nav/AppNavigator.js
  22 files around src/selectors.js
  10 files around src/types.js
   6 files around src/webview/MessageList.js
   4 files around src/api/apiTypes.js