UserStatusScreen: Move and re-style "clear" button
This doesn't yet change the main behavior of the button when
pressed, i.e., we still clear the inputs *and* tell the server that
we want to unset the status.

Soon, we'll have it match the web app by only clearing the inputs.
Then the user can apply the change with the submit button, which
we'll rename from "Update" to "Save".