UserStatusScreen [nfc]: Consistently represent "unset" in statusText state
We've *said* statusText is `string | null` (changed from
`string | void` in our recent model rework, 2106381e4).

I would like to make it `string`, to match the `value` prop of the
`Input` component type. That seems like a responsible thing for
UserStatusScreen to do, since it's using the Input as a controlled
input: https://reactjs.org/docs/forms.html#controlled-components

Before we do that, we should at least fix this inconsistency, where
the value of the input corresponds to "unset" (empty string) but
statusText doesn't represent "unset" the way it says it will (null).

It hasn't been a live bug, and this is NFC, because it just so
happens that api.updateUserStatus converts to the empty string
anyway, to talk to the server.