UserStatusScreen: Make status suggestions use new emoji-status feature!
When available, that is. Support was added in FL 86.

If emoji statuses are supported, we use the suggested emoji for the
emoji status, and the suggested text for the text status.

If emoji statuses aren't supported…we do end up just cutting the
emoji out of the status suggestion, which I guess is a small
downside. For these older servers, we could have kept up what we do
now: stick the status emoji at the start of the suggested text. But
I think achieving that has a higher complexity cost than it's worth.
I've never really used this UI, I'm not sure how common it is to do
so, and people are still free to depart from the suggested text,
including by adding emojis, by using the text input.