api types: Fix propagate_mode param on updateMessage
This is an enum, the same one found on update_message events;
it's not a boolean.

This isn't a matter of being out of date: there was never a server
version where `propagate_mode` was a boolean.  It's been this enum
since it was introduced, in zulip/zulip@c588c7938 in 2013.  (It
replaced a boolean with a different name, `propagate_subject`.)

The main reason we've never noticed is that we don't yet have any code
to make use of this feature.