ComposeBox [nfc]: Inline handleMessageChange in setMessageInputValue
After all, this code isn't *handling* a change that was made in the
input value. It's causing a change.

Also, we'd like to change `setMessageInputValue`'s use of
`setState`, next, so that its callers can pass a function that
computes the new state from the state at the time the update is to
be applied. Best to get `handleMessageChange` out of the way before
we do that; we don't have a reason to make that kind of change
there.