msglist tests: Fix flake; make test data more representative
One of the `check` functions hasn't been using our `baseState` that
we put together with stream1 and stream2, to give background for the
messages in those streams (streamMessages1, streamMessages2, etc.)
that we pass to `check`.

So, use that `baseState`.

Also, the question of whether those streams are subscribed has
effectively been left up to chance: if `eg.plusReduxState` *happens*
to have a `subscriptions` entry for stream1 or stream2 (i.e., one
where the stream_id matches), then we render a stream header as
"subscribed"; otherwise not [1]. `eg.plusReduxState` currently has
two `subscriptions` entries, both with random stream_ids. Most of
the time they differ from stream1 and stream2 (effectively making
those streams "unsubscribed"), but we saw a flake in CI for #5290
where they must've collided.

Settle the question by saying the streams are actually subscribed,
which is the common case.

[1] If "subscribed", the background color is the stream's color
    (#123456 in current example data); otherwise, the fallback
    hsl(0, 0%, 80%). And the text color is
    `foregroundColorFromBackground` of that.