user [nfc]: Drop unused features of filterUserList
We haven't used the filter-out-self feature here since 72acf9ef9,
in 2017.

And without that, it would never make sense to not pass a filter
string, because then the function would be doing nothing at all.
In fact, given that the self-user parameter comes after the filter
string, this means it could never have made sense to not pass a
filter string.