storage tests: Apply avatar "wakeup" to places users hide in state
The users come in several places in the initial data from the server.
We should probably combine all three of them up front into one place
in the state, but at present we don't (we do so in a caching selector
instead.)  So this small hack in a test should apply to all three.