SmartUrlInput: Remove workaround for an RN bug that's less bothersome now
Less bothersome because, after the previous commit, the only
remaining .focus() call is for convenience (autofocus when
back-navigating to the screen), not necessity.

With this RN bug, whether we work around it or not, it remains
possible to focus a TextInput by tapping on it.

So why "necessity"? Before the previous commit, we had a hack where
text elements were trying to blend in as part of a TextInput, with
the real TextInput being as little as 1px wide when empty. So,
pressing the text elements was an important way to give the input
focus, and that was achieved with .focus().