flow-typed libdef: Use instance not class type in Formatters.getFoo return
Here's Flow's built-in definition for the `Intl` object [1]:

  declare var Intl: {
    Collator: Class<Intl$Collator>,
    DateTimeFormat: Class<Intl$DateTimeFormat>,
    NumberFormat: Class<Intl$NumberFormat>,
    PluralRules: ?Class<Intl$PluralRules>,
    getCanonicalLocales?: (locales?: Intl$Locales) => Intl$Locale[],
    ...
  }

(It uses Flow's built-in `Class` type:
https://flow.org/en/docs/types/utilities/#toc-class )

So, for Intl, the names with $ stand for instance types, and the
names with . stand for class types.

It looks like the touched lines of code really want to say something
like

   getDateTimeFormat: ConstructorOf<Intl.DateTimeFormat>

, where `ConstructorOf` is a name I just made up, and Flow doesn't
yet have a tool that would fit that name; see
  https://github.com/facebook/flow/issues/1409#issuecomment-184443566

Without that, it's at least easy to preserve the return value of
each of these functions. As with any constructor function, the
return value is an instance object, not the class object, which is
why we make this change.

[1] https://github.com/facebook/flow/blob/v0.141.0/lib/intl.js#L8-L15