event: List all remaining documented events, and explicitly ignore
With comments about in what circumstances we'll need to start
handling them.

Generated the list automatically from the API metadata, like so:

  $ python -c 'if 1:
        import yaml, json, sys
        z = yaml.safe_load(open("../zulip/zerver/openapi/zulip.yaml"))
        json.dump(z, sys.stdout)
      ' | jq '
        .paths["/events"].get.responses["200"].content["application/json"]
          .schema.allOf[2].properties.events.items.oneOf
          [] | ((.properties // .allOf[1].properties).type.allOf[1].enum[0] // .)
      ' -r \
    | sort -u \
    | perl -lne 'print "  $_: null,"'

(That `jq` formula is a bit brittle to the details of how our
openapi/zulip.yaml is written.  If you're trying it in the future and
that has changed in a way that breaks this, the `// .` should cause
the whole objects affected to get spit out.  In that case, delete the
`| sort -u` to see more clearly for debugging.)