react-intl libdef [nfc]: Say "Omit" instead of "Exclude"
This doesn't actually improve Flow coverage or change any types;
both Omit and Exclude are utility types in TypeScript that Flow
doesn't have. But in the original TypeScript, the use of Exclude for
these was a mistake; they should have been Omit; see discussion at
  https://github.com/zulip/zulip-mobile/pull/5308#issuecomment-1077066675

We're about to go through all uses of Omit and replace them with
something that works; these will then be fixed.