netinfo libdef: Inline helper NetInfoConnectedDetails
NetInfoConnectedDetails is inexact; so, make the object type that it
contributes to (at NetInfoConnectedState.details) inexact.

Also, the type param D is often inexact; that's another reason to
make the outer object inexact.

As a consequence, Flow v0.149 would rightly complain about possibly
trying to `logging.warn` non-JSONable contents of the object. So,
comment about that.