event: Explicitly check for RequestError before using its properties
The reason we're looking at an `httpStatus` property here is that
we expect the error to be one of our `RequestError` objects.  If
it's some other unexpected type of object, and yet happens to have
a property named `httpStatus`, then we don't know what the
semantics of that property is and can't reliably infer anything
from inspecting it.