types: Explicitly recover `catch` type-checking in various places
Due to facebook/flow#2470, Flow on its own will let a `catch` body
implicitly assume absolutely anything it likes about the caught
exception -- it gives it the type `empty`.

With an extra line of code, we can instead revert to the truth, which
is that the exception could be anything at all -- the code can't
reliably assume anything at all (short of auditing the whole potential
call tree under the `try`), or in other words the type is `mixed`.

We already do this in a handful of places.  Do it in eight more of
them, aiming for all the spots where we do anything interesting with
the caught exception.  This means that when we do things like inspect
particular properties of the exception, Flow is now able to verify as
usual that the assumptions we make are correctly justified, e.g. by
`instanceof` checks.

Previous commits fixed a couple of spots where our assumptions
*weren't* in fact so justified.  In this commit, we make no logic
changes; we do add some casts to make assumptions explicit that were
already present, and relax some downstream expectations about error
objects, where assumptions like `error: Error` weren't actually used.