api types: Stop using FixmeUntypedFetchResult where inappropriate
And just use $FlowFixMe instead.

These aren't places where we're dealing with a `fetch` return value,
which is what FixmeUntypedFetchResult is for. The `apiCall` function
does do a `fetch`, but it doesn't return the result directly; it
does some processing of it.

It's not clear what exactly these types should be, but that's
because the interesting parts of these values are highly specific to
the caller. For now, make a note on `apiCall` about that, with a
$FlowFixMe, and let the $FlowFixMe propagate to `apiCall`'s callers.

See discussion:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/FixmeUntypedFetchResult/near/1363106