emoji tests: Add tests for differences from shared emoji matcher
These test differences in behavior I spotted from studying the
shared code that does emoji matches, and comparing with our code.

One is for some behavior that actually feels pretty broken in shared
(visible in the web app by trying to type an emoji in the compose box,
with the `:foo:` syntax), but works well in mobile.  ... OK, I've gone
and fixed that in web/shared: https://github.com/zulip/zulip/pull/21871

The other is for some behavior that the shared code gets right, and
would be a pretty nice UI improvement in mobile for the add-reaction
screen.  So that test is skipped for now, to be enabled shortly when
we switch to using the shared code here.