shared: Upgrade to 0.0.14
This brings a series of improvements to the emoji typeahead code:
  https://github.com/zulip/zulip/pull/21871

In particular, for deciding which emoji match the query, the shared
code had had some regressions (as well as improvements) compared to
our mobile version, and we hadn't yet switched to using it.  These
changes fix those, so we can now switch to shared there without
regressions.