deps: Temporarily use fork for @react-navigation/bottom-tabs
To get chrisbobbe/react-navigation@b2343cb25, which I've published
to NPM in @zulip/react-navigation-bottom-tabs@5.11.16-0.zulip.1.

It cherry-picks react-navigation/react-navigation@d87857e5d, which
is a compatibility fix for RN v0.65 that React Navigation published
in v6 but didn't backport to v5.

So this fix will become unnecessary when we're on React Navigation
6; that's issue #4936.

A small handful of things went wrong when building the .d.ts files
for the publish to NPM, so some TypeScript types are lost, which is
fine because we don't use TypeScript. But the runtime code looks
like it's updated in the ways we'd want and expect.