mute: Enrich model to support new user_topic API
Ideally we'd rename this to something like `state.userTopic` too.
But it's currently a pain to actually successfully rename a state
subtree (and not have its ghost return at the old name): #5381.
So just keep calling it "mute" for now.

Also add tests for the model.  Note that we don't have to update
any of the existing tests!  They use `makeMuteState`, which uses
the reducer; so they get the new data structure automatically.