android: Work around issue with "open failed: EACCES (Permission denied)"
See
  https://github.com/facebook/react-native/issues/32735#issue-1077061487
, where this workaround is copied from, and see a later comment on
the same issue saying that the official fix will arrive in RN v0.68.

Using Android Studio, Greg and I found that the version of soloader
that was pulled in before was 0.10.1. (Under the Packages list, we
went to Libraries/com/facebook/soloader, opened a random file, and
saw the version in Android Studio's title bar.)