nav types [nfc]: Replace NavigationContainerType with typeof NavigationContainer
Upstream doesn't give a name to the type of this value it exports.
But we can take `typeof` of it.  And because the only exported name
that refers to is the name of a value -- and the two type definitions
must agree about what names refer to values, because that's part of
the runtime reality that they're both just seeking to describe --
that approach works already with our existing type definitions.