EditStreamCard: Use new getOwnUserRole and roleIsAtLeast
We already have checks for making web-public streams here, and we'd
like to add more checks, for regular public and private streams.
We'll want those checks to work on pre-4.0 servers, which don't send
a `role` property on users. So, use getOwnUserRole. See there for
how it gives pre-4.0 support.

(We don't have a compatibility problem like that with the existing
web-public-stream checks because we never give the option to make
web-public streams when the server is pre-4.0.)

The `roleIsAtLeast` change should be NFC, given the current set of
roles, but makes the code robust to adding more roles above Admin,
in case we decide to do that.